:recycle: refactor : 연관관계 수정 , id 타입 변경, baseEntity 추가

+++ hugme-dog/src/main/java/spring/hugme/domain/magazine/entity/Resource.java
@@ -18,11 +18,13 @@ import spring.hugme.infra.entity.BaseEntity;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
+//하나의 매거진에 여러개의 리소스가 들어갈수 있고 하나의 리소스는 여러개의 매거진이 참조 될수 있다
+//ManyToMany이므로 중간다리 역할할 하나의 엔티티가 더 필요하다 !!
 public class Resource extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int resourceId;
+  private Long resourceId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "category_id")
